requirements.txt
Django==4.1.1
black==22.8.0
gunicorn==20.1.0
dj-database-url==0.5.0
django-encrypted-id==0.3.2